Rename mjs files to js files
